+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.72] - 2024-09-25
+
 ### Add
 
 - Attraction component
@@ -455,7 +457,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Several UI fixes in next event, and add price format
 
-[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.71...HEAD
+[unreleased]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.72...HEAD
+[0.0.72]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.71...0.0.72
 [0.0.71]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.70...0.0.71
 [0.0.70]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.69...0.0.70
 [0.0.69]: https://github.com/kbzaso/5lc-sveltkit-sanity/compare/0.0.68...0.0.69